feat: added component
